Added the console-service (with color-codes), added io-port-permissions and some other stuff

+++ kernel/h/vfs.h
@@ -36,7 +36,7 @@ struct sVFSNode {
 	union {
 		struct {
 			fRead readHandler;
-			u16 size;
+			u32 size;
 			void *cache;
 		} info;
 		sProc *proc;
